basic idae of actor prep
