+++ src/module/actor/SR5Actor.ts
@@ -17,6 +17,8 @@ import LimitField = Shadowrun.LimitField;
 import { SYSTEM_NAME } from '../constants';
 import SR5ActorData = Shadowrun.SR5ActorData;
 import { SR5ItemDataWrapper } from '../item/SR5ItemDataWrapper';
+import { BaseActorPrep } from './prep/BaseActorPrep';
+import SR5ActorType = Shadowrun.SR5ActorType;
 
 export class SR5Actor extends Actor {
     async update(data, options?) {
@@ -46,389 +48,23 @@ export class SR5Actor extends Actor {
         }
     }
 
-    /**
-     * Prepare Matrix data on the actor
-     * - if an item is equipped, it will use that data
-     * - if it isn't and player is technomancer, it will use that data
-     * @param data
-     * @param items
-     */
-    private static prepareMatrix(data: SR5ActorData, items: SR5ItemDataWrapper[]) {
-        const { matrix, attributes, limits } = data;
-
-        // clear matrix data to defaults
-        matrix.firewall.value = Helpers.totalMods(matrix.firewall.mod);
-        matrix.data_processing.value = Helpers.totalMods(matrix.data_processing.mod);
-        matrix.attack.value = Helpers.totalMods(matrix.attack.mod);
-        matrix.sleaze.value = Helpers.totalMods(matrix.sleaze.mod);
-        matrix.condition_monitor.max = 0;
-        matrix.rating = 0;
-        matrix.name = '';
-        matrix.device = '';
-
-        // get the first equipped device, we don't care if they have more equipped -- it shouldn't happen
-        const device = items.find((item) => item.isEquipped() && item.isDevice());
-
-        if (device) {
-            const conditionMonitor = device.getConditionMonitor();
-            matrix.device = device.getId();
-            matrix.condition_monitor.max = conditionMonitor.max;
-            matrix.condition_monitor.value = conditionMonitor.value;
-            matrix.rating = device.getRating();
-            matrix.is_cyberdeck = device.isCyberdeck();
-            matrix.name = device.getName();
-            matrix.item = device.getData();
-            const deviceAtts = device.getASDF();
-            if (deviceAtts) {
-                // setup the actual matrix attributes for the actor
-                for (const [key, value] of Object.entries(deviceAtts)) {
-                    if (value && matrix[key]) {
-                        matrix[key].value += value.value;
-                        matrix[key].device_att = value.device_att;
-                    }
-                }
-            }
-        } // if we don't have a device, use living persona
-        else if (data.special === 'resonance') {
-            matrix.firewall.value += Helpers.calcTotal(attributes.willpower);
-            matrix.data_processing.value += Helpers.calcTotal(attributes.logic);
-            matrix.rating = Helpers.calcTotal(attributes.resonance);
-            matrix.attack.value += Helpers.calcTotal(attributes.charisma);
-            matrix.sleaze.value += Helpers.calcTotal(attributes.intuition);
-            matrix.name = game.i18n.localize('SR5.LivingPersona');
-        }
-
-        // set matrix condition monitor to max if greater than
-        if (matrix.condition_monitor.value > matrix.condition_monitor.max) {
-            matrix.condition_monitor.value = matrix.condition_monitor.max;
-        }
-
-        // add matrix attributes to both limits and attributes as hidden entries
-        ['firewall', 'sleaze', 'data_processing', 'firewall'].forEach((key) => {
-            if (matrix[key]) {
-                const label = CONFIG.SR5.matrixAttributes[key];
-                const { value, base, mod } = matrix[key];
-                const hidden = true;
-
-                limits[key] = {
-                    value,
-                    base,
-                    mod,
-                    label,
-                    hidden,
-                };
-                attributes[key] = {
-                    value,
-                    base,
-                    mod,
-                    label,
-                    hidden,
-                };
-            }
-        });
-    }
-
-    /**
-     * Prepare the armor data for the Item
-     * - will only allow one "Base" armor item to be used
-     * - all "accessories" will be added to the armor
-     * @param data
-     * @param items
-     */
-    private static prepareArmor(data: SR5ActorData, items: SR5ItemDataWrapper[]) {
-        const { armor } = data;
-        armor.base = 0;
-        armor.value = 0;
-        armor.mod = {};
-        for (const element of Object.keys(CONFIG.SR5.elementTypes)) {
-            armor[element] = 0;
-        }
-
-        const equippedArmor = items.filter((item) => item.isArmor() && item.isEquipped());
-        equippedArmor?.forEach((item) => {
-            if (item.isArmorAccessory()) {
-                armor.mod[item.getName()] = item.getArmorValue();
-            } // if not a mod, set armor.value to the items value
-            else {
-                armor.base = item.getArmorValue();
-                armor.label = item.getName();
-                for (const element of Object.keys(CONFIG.SR5.elementTypes)) {
-                    armor[element] = item.getArmorElements()[element];
-                }
-            }
-        });
-
-        if (data.modifiers['armor']) armor.mod[game.i18n.localize('SR5.Bonus')] = data.modifiers['armor'];
-        // SET ARMOR
-        armor.value = armor.base + Helpers.totalMods(armor.mod);
-    }
-
-    /**
-     * Prepare actor data for cyberware changes
-     * - this calculates the actors essence
-     * @param data
-     * @param items
-     */
-    private static prepareCyberware(data: SR5ActorData, items: SR5ItemDataWrapper[]) {
-        const { attributes } = data;
-        let totalEssence = 6;
-        items
-            .filter((item) => item.isCyberware() && item.isEquipped())
-            .forEach((item) => {
-                if (item.getEssenceLoss()) {
-                    totalEssence -= item.getEssenceLoss();
-                }
-            });
-        attributes.essence.value = +(totalEssence + Number(data.modifiers['essence'])).toFixed(3);
-    }
-
-    /**
-     * Prepare actor data for attributes
-     * @param data
-     */
-    private static prepareAttributes(data: SR5ActorData) {
-        const { attributes } = data;
-
-        // set the value for the attributes
-        for (let [key, attribute] of Object.entries(attributes)) {
-            Helpers.calcTotal(attribute);
-            // add labels
-            attribute.label = CONFIG.SR5.attributes[key];
-        }
-    }
-
-    /**
-     * Prepare actor data for skills
-     * @param data
-     */
-    private static prepareSkills(data: SR5ActorData) {
-        const { language, active, knowledge } = data.skills;
-        if (language) {
-            if (!language.value) language.value = {};
-            language.attribute = 'intuition';
-        }
-
-        // function that will set the total of a skill correctly
-        const prepareSkill = (skill) => {
-            skill.mod = {};
-            if (!skill.base) skill.base = 0;
-            if (skill.bonus?.length) {
-                for (let bonus of skill.bonus) {
-                    skill.mod[bonus.key] = bonus.value;
-                }
-            }
-            Helpers.calcTotal(skill);
-        };
-
-        // setup active skills
-        for (const skill of Object.values(active)) {
-            if (!skill.hidden) {
-                prepareSkill(skill);
-            }
-        }
-
-        const entries = Object.entries(data.skills.language.value);
-        // remove entries which are deleted TODO figure out how to delete these from the data
-        entries.forEach(([key, val]: [string, { _delete?: boolean }]) => val._delete && delete data.skills.language.value[key]);
-
-        for (let skill of Object.values(language.value)) {
-            prepareSkill(skill);
-            skill.attribute = 'intuition';
-        }
-
-        // setup knowledge skills
-        for (let [, group] of Object.entries(knowledge)) {
-            const entries = Object.entries(group.value);
-            // remove entries which are deleted TODO figure out how to delete these from the data
-            group.value = entries
-                .filter(([, val]) => !val._delete)
-                .reduce((acc, [id, skill]) => {
-                    prepareSkill(skill);
-
-                    // set the attribute on the skill
-                    skill.attribute = group.attribute;
-                    acc[id] = skill;
-                    return acc;
-                }, {});
-        }
-
-        // skill labels
-        for (let [skillKey, skillValue] of Object.entries(active)) {
-            skillValue.label = CONFIG.SR5.activeSkills[skillKey];
-        }
-    }
-
-    /**
-     * Prepare the actor data limits
-     * @param data
-     */
-    private static prepareLimits(data: SR5ActorData) {
-        const { limits, attributes, modifiers } = data;
-
-        // SETUP LIMITS
-        limits.physical.value =
-            Math.ceil((2 * attributes.strength.value + attributes.body.value + attributes.reaction.value) / 3) + Number(modifiers['physical_limit']);
-        limits.mental.value =
-            Math.ceil((2 * attributes.logic.value + attributes.intuition.value + attributes.willpower.value) / 3) + Number(modifiers['mental_limit']);
-        limits.social.value =
-            Math.ceil((2 * attributes.charisma.value + attributes.willpower.value + attributes.essence.value) / 3) + Number(modifiers['social_limit']);
-
-        // limit labels
-        for (let [limitKey, limitValue] of Object.entries(limits)) {
-            limitValue.label = CONFIG.SR5.limits[limitKey];
-        }
-    }
-
-    /**
-     * Prepare actor data condition monitors (aka Tracks)
-     * @param data
-     */
-    private static prepareConditionMonitors(data: SR5ActorData) {
-        const { track, attributes, modifiers } = data;
-
-        // TODO we will have grunts eventually that only have one track
-        track.physical.max = 8 + Math.ceil(attributes.body.value / 2) + Number(modifiers['physical_track']);
-        track.physical.overflow.max = attributes.body.value;
-        track.stun.max = 8 + Math.ceil(attributes.willpower.value / 2) + Number(modifiers['stun_track']);
-
-        // tracks
-        for (let [trackKey, trackValue] of Object.entries(track)) {
-            trackValue.label = CONFIG.SR5.damageTypes[trackKey];
-        }
-    }
-
-    /**
-     * Prepare actor data movement
-     * @param data
-     */
-    private static prepareMovement(data: SR5ActorData) {
-        const { attributes, modifiers } = data;
-        const movement = data.movement;
-        // default movement: WALK = AGI * 2, RUN = AGI * 4
-        movement.walk.value = attributes.agility.value * (2 + Number(modifiers['walk']));
-        movement.run.value = attributes.agility.value * (4 + Number(modifiers['run']));
-    }
-
-    /**
-     * Prepare the modifiers that are displayed in the Misc. tab
-     * @param data
-     */
-    private static prepareModifiers(data: SR5ActorData) {
-        if (!data.modifiers) data.modifiers = {};
-        const modifiers = {};
-        let miscTabModifiers = [
-            'soak',
-            'drain',
-            'armor',
-            'physical_limit',
-            'social_limit',
-            'mental_limit',
-            'stun_track',
-            'physical_track',
-            'meat_initiative',
-            'meat_initiative_dice',
-            'astral_initiative',
-            'astral_initiative_dice',
-            'matrix_initiative',
-            'matrix_initiative_dice',
-            'composure',
-            'lift_carry',
-            'judge_intentions',
-            'memory',
-            'walk',
-            'run',
-            'defense',
-            'wound_tolerance',
-            'essence',
-            'fade',
-        ];
-        miscTabModifiers.sort();
-        // force global to the top
-        miscTabModifiers.unshift('global');
-
-        for (let item of miscTabModifiers) {
-            modifiers[item] = Number(data.modifiers[item]) || 0;
-        }
-
-        data.modifiers = modifiers;
-    }
-
-    /**
-     * Prepare actor data for initiative
-     * @param data
-     */
-    private static prepareInitiative(data: SR5ActorData) {
-        const { initiative, attributes, modifiers, matrix } = data;
-        initiative.meatspace.base.base = attributes.intuition.value + attributes.reaction.value + Number(modifiers['meat_initiative']);
-        initiative.meatspace.dice.base = 1 + Number(modifiers['meat_initiative_dice']);
-        initiative.astral.base.base = attributes.intuition.value * 2 + Number(modifiers['astral_initiative']);
-        initiative.astral.dice.base = 2 + Number(modifiers['astral_initiative_dice']);
-        initiative.matrix.base.base = attributes.intuition.value + data.matrix.data_processing.value + Number(modifiers['matrix_initiative']);
-        initiative.matrix.dice.base = matrix.hot_sim ? 4 : 3 + Number(modifiers['matrix_initiative_dice']);
-        if (initiative.perception === 'matrix') initiative.current = initiative.matrix;
-        else if (initiative.perception === 'astral') initiative.current = initiative.astral;
-        else {
-            initiative.current = initiative.meatspace;
-            initiative.perception = 'meatspace';
-        }
-        initiative.current.dice.value = initiative.current.dice.base;
-        if (initiative.edge) initiative.current.dice.value = 5;
-        initiative.current.dice.value = Math.min(5, initiative.current.dice.value); // maximum of 5d6 for initiative
-        initiative.current.dice.text = `${initiative.current.dice.value}d6`;
-        initiative.current.base.value = initiative.current.base.base;
-    }
-
-    /**
-     * Prepare actor data for wounds
-     * @param data
-     */
-    private static prepareWounds(data: SR5ActorData) {
-        const { modifiers, track } = data;
-        const count = 3 + Number(modifiers['wound_tolerance']);
-        const stunWounds = Math.floor(data.track.stun.value / count);
-        const physicalWounds = Math.floor(data.track.physical.value / count);
-
-        track.stun.wounds = stunWounds;
-        track.physical.wounds = physicalWounds;
-
-        data.wounds = {
-            value: stunWounds + physicalWounds,
-        };
-    }
-
     prepareData() {
         super.prepareData();
 
-        const actorData = this.data;
-        // @ts-ignore
-        const items: SR5ItemDataWrapper[] = actorData.items.map((item) => new SR5ItemDataWrapper(item));
+        const actorData = this.data as SR5ActorType;
+        const prepper = new BaseActorPrep(actorData);
+        prepper.prepareModifiers();
+        prepper.prepareArmor();
+        prepper.prepareCyberware();
+        prepper.prepareSkills();
+        prepper.prepareAttributes();
+        prepper.prepareMatrix();
+        prepper.prepareLimits();
+        prepper.prepareConditionMonitors();
+        prepper.prepareMovement();
+        prepper.prepareWounds();
+
         const data = actorData.data;
-        const { attributes }: { attributes: Attributes } = data;
-
-        attributes.magic.hidden = !(data.special === 'magic');
-        attributes.resonance.hidden = !(data.special === 'resonance');
-
-        // PARSE WEAPONS AND SET VALUES AS NEEDED
-
-        // modifiers must be handled first
-        SR5Actor.prepareModifiers(data);
-        // items preparations
-        SR5Actor.prepareArmor(data, items);
-        SR5Actor.prepareCyberware(data, items);
-        SR5Actor.prepareSkills(data);
-        SR5Actor.prepareAttributes(data);
-        // matrix prep must happen after attributes
-        SR5Actor.prepareMatrix(data, items);
-        // limits prep must happen after attributes
-        SR5Actor.prepareLimits(data);
-        SR5Actor.prepareConditionMonitors(data);
-        SR5Actor.prepareMovement(data);
-        SR5Actor.prepareInitiative(data);
-        SR5Actor.prepareWounds(data);
-
-        // CALCULATE RECOIL
-        data.recoil_compensation = 1 + Math.ceil(attributes.strength.value / 3);
-
         if (data.magic.drain && !data.magic.drain.mod) data.magic.drain.mod = {};
     }
