+++ src/module/types/actor/ShadowrunCharacter.ts
@@ -1,4 +1,5 @@
 /// <reference path="../Shadowrun.ts" />
+
 declare namespace Shadowrun {
     export type SR5ActorSheetData = ActorSheetData & {
         data: SR5ActorData;
@@ -7,6 +8,19 @@ declare namespace Shadowrun {
 
     export type SR5SheetFilters = {
         skills: string;
+    };
+
+    export type SR5ActorType = ActorData & {
+        name: string;
+        _id: string;
+        folder: string | null;
+        type: string;
+        data: SR5ActorData;
+        items: Collection<Item>;
+        flags: object;
+        permission: {
+            default: string;
+        };
     };
 
     export type SR5ActorData = ActorData & {
@@ -26,11 +40,12 @@ declare namespace Shadowrun {
         armor: ActorArmor;
         initiative: Initiative;
         wounds: WoundType;
+        recoil_compensation: number;
     };
 
     export type WoundType = {
         value: number;
-    }
+    };
 
     export type Modifiers = {
         [name: string]: NumberOrEmpty;
@@ -38,9 +53,10 @@ declare namespace Shadowrun {
 
     export type InitiativeType = {
         base: BaseValuePair<number> & ModifiableValue;
-        dice: BaseValuePair<number> & ModifiableValue & {
-            text: string;
-        };
+        dice: BaseValuePair<number> &
+            ModifiableValue & {
+                text: string;
+            };
     };
 
     export type Initiative = {
